requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
dnspython==2.7.0
faiss-cpu==1.11.0
filelock==3.18.0
fsspec==2025.5.0
hf-xet==1.1.2
huggingface-hub==0.14.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
packaging==25.0
pillow==11.2.1
pymongo==4.13.0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==2.2.2
sentencepiece==0.2.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.13.3
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.30.2
typing_extensions==4.13.2
urllib3==2.4.0
